requirements.txt
fastapi
python-dotenv
openai
jira
pydantic
sqlalchemy
uvicorn
python-telegram-bot
